refactor: restructure of controllers and replaced placeholders with dynamic data in header
