+++ client/src/main/resources/templates/fragments/assets/navbar.html
@@ -1,15 +1,21 @@
-<section class="flex flex-col lg:flex-row justify-between items-center mb-2 px-6" th:fragment="header">
+<section class="flex flex-col lg:flex-row justify-between items-center mb-8 px-6" th:fragment="header">
     <div class="mb-4 lg:mb-0">
         <h2 class="text-lg font-bold">Hola %placeholderUsu%, bienvenido</h2>
         <h3 class="text-4xl font-bold text-[#244444] mt-1" th:text="${pageTitle}">%placeholderPage%</h3>
     </div>
 
 
-    <div class="flex items-center space-x-2 sm:hidden">
+    <div class="flex items-center space-x-2">
         <div class="text-right">
-            <span class="font-bold">%placeholderUsu%</span><br>
-            <span class="text-sm text-gray-500">%placeholderRol%</span>
+            <span class="font-bold hidden md:inline-block"><span class="capitalize" th:text="${user.username}">%placeholderUsu%</span></span><br>
+            <span class="text-sm text-gray-500 hidden md:inline-block" th:switch="${user.role.id}">
+                <p th:case="2">Jefe de Operaciones</p>
+                <p th:case="3">Operario</p>
+            </span>
         </div>
-        <div class="w-10 h-10 bg-gray-300 rounded-full"></div>
+        <div class="w-10 h-10 bg-gray-300 rounded-full hidden sm:block" th:switch="${user.role.id}">
+            <img th:case="2" th:src="@{/img/BOSS.png}" alt="Jefe de Operaciones - PFP" class="mx-auto rounded-full">
+            <img th:case="3" th:src="@{/img/EMPLOYEE.png}" alt="Operario - PFP" class="mx-auto rounded-full">
+        </div>
     </div>
 </section>
